src/requirements.txt
Pillow==2.2.1